split selector and hooks out from groupEntries component and add unit tests
