+++ app/packages/core/src/components/Sidebar/Entries/GroupEntries.tsx
@@ -10,8 +10,6 @@ import VisibilityIcon from "@mui/icons-material/Visibility";
 import React, { useRef, useState } from "react";
 import {
   SetterOrUpdater,
-  selectorFamily,
-  useRecoilCallback,
   useRecoilStateLoadable,
   useRecoilValue,
   useRecoilValueLoadable,
@@ -19,250 +17,21 @@ import {
 import styled from "styled-components";
 
 import * as fos from "@fiftyone/state";
-import { removeKeys } from "@fiftyone/utilities";
 
 import { PillButton, useTheme } from "@fiftyone/components";
-import { datasetName } from "@fiftyone/state";
 import {
-  attributeVisibility,
-  modalAttributeVisibility,
-} from "@fiftyone/state/src/recoil/attributeVisibility";
+  useClearActive,
+  useClearFiltered,
+  useClearVisibility,
+  useDeleteGroup,
+  useRenameGroup,
+} from "@fiftyone/state/src/hooks/useGroupEntries";
+import {
+  numGroupFieldsActive,
+  numGroupFieldsFiltered,
+  numGroupFieldsVisible,
+} from "@fiftyone/state/src/recoil/groupEntries";
 import Draggable from "./Draggable";
-
-const groupLength = selectorFamily<number, { modal: boolean; group: string }>({
-  key: "groupLength",
-  get:
-    (params) =>
-    ({ get }) =>
-      get(fos.sidebarGroup({ ...params, loading: true })).length,
-});
-
-const numGroupFieldsFiltered = selectorFamily<
-  number,
-  { modal: boolean; group: string }
->({
-  key: "numGroupFieldsFiltered",
-  get:
-    (params) =>
-    ({ get }) => {
-      let count = 0;
-
-      let f = null;
-
-      if (params.modal) {
-        const labels = get(fos.labelPaths({ expanded: false }));
-        f = (path) => labels.includes(path);
-      }
-
-      for (const path of get(fos.sidebarGroup({ ...params, loading: true }))) {
-        if (
-          get(fos.fieldIsFiltered({ path, modal: params.modal })) &&
-          (!f || f(path))
-        )
-          count++;
-      }
-
-      return count;
-    },
-});
-
-const numGroupFieldsVisible = selectorFamily<
-  number,
-  { modal: boolean; group: string }
->({
-  key: "numGroupFieldsVisible",
-  get:
-    (params) =>
-    ({ get }) => {
-      let count = 0;
-
-      let f = null;
-
-      if (params.modal) {
-        const labels = get(fos.labelPaths({ expanded: false }));
-        f = (path) => labels.includes(path);
-      }
-
-      for (const path of get(fos.sidebarGroup({ ...params, loading: true }))) {
-        if (
-          get(fos.fieldHasVisibilitySetting({ path, modal: params.modal })) &&
-          (!f || f(path))
-        )
-          count++;
-      }
-
-      return count;
-    },
-});
-
-const numGroupFieldsActive = selectorFamily<
-  number,
-  { modal: boolean; group: string }
->({
-  key: "numGroupFieldsActive",
-  get:
-    (params) =>
-    ({ get }) => {
-      let active = get(fos.activeFields({ modal: params.modal }));
-
-      let f = null;
-
-      if (params.modal) {
-        const labels = get(fos.labelPaths({ expanded: false }));
-        f = (path) => labels.includes(path);
-        active = active.filter((p) => f(p));
-      }
-
-      f = get(fos.textFilter(params.modal));
-
-      if (params.group === "tags") {
-        return active.filter(
-          (p) => p.startsWith("tags.") && p.slice("tags.".length).includes(f)
-        ).length;
-      }
-
-      if (params.group === "tags") {
-        return active.filter(
-          (p) =>
-            p.startsWith("_label_tags.") &&
-            p.slice("_label_tags.".length).includes(f)
-        ).length;
-      }
-
-      const paths = new Set(
-        get(fos.sidebarGroup({ ...params, loading: true }))
-      );
-
-      return active.filter((p) => p.includes(f) && paths.has(p)).length;
-    },
-});
-
-export const replace = {};
-
-export const useRenameGroup = (modal: boolean, group: string) => {
-  return useRecoilCallback(
-    ({ set, snapshot }) =>
-      async (newName: string) => {
-        newName = newName.toLowerCase();
-
-        const current = await snapshot.getPromise(
-          fos.sidebarGroupsDefinition(modal)
-        );
-        if (
-          !fos.validateGroupName(
-            current.map(({ name }) => name).filter((name) => name !== group),
-            newName
-          )
-        ) {
-          return false;
-        }
-
-        const newGroups = current.map(({ name, ...rest }) => ({
-          name: name === group ? newName : name,
-          ...rest,
-        }));
-
-        const view = await snapshot.getPromise(fos.view);
-        const shown = await snapshot.getPromise(
-          fos.groupShown({ modal, group, loading: true })
-        );
-
-        replace[newName] = group;
-
-        set(fos.groupShown({ group: newName, modal, loading: true }), shown);
-        set(fos.sidebarGroupsDefinition(modal), newGroups);
-        !modal &&
-          fos.persistSidebarGroups({
-            dataset: await snapshot.getPromise(datasetName),
-            stages: view,
-            sidebarGroups: newGroups,
-          });
-        return true;
-      },
-    []
-  );
-};
-
-export const useDeleteGroup = (modal: boolean, group: string) => {
-  const numFields = useRecoilValue(groupLength({ modal, group }));
-  const onDelete = useRecoilCallback(
-    ({ set, snapshot }) =>
-      async () => {
-        const groups = await snapshot.getPromise(
-          fos.sidebarGroups({ modal, loading: true })
-        );
-        set(
-          fos.sidebarGroups({ modal, loading: true }),
-          groups.filter(({ name }) => name !== group)
-        );
-      },
-    []
-  );
-
-  if (numFields) {
-    return null;
-  }
-
-  return onDelete;
-};
-
-const useClearActive = (modal: boolean, group: string) => {
-  return useRecoilCallback(
-    ({ set, snapshot }) =>
-      async () => {
-        const paths = await snapshot.getPromise(
-          fos.sidebarGroup({ modal, group, loading: true })
-        );
-        const active = await snapshot.getPromise(fos.activeFields({ modal }));
-
-        set(
-          fos.activeFields({ modal }),
-          active.filter((p) => !paths.includes(p))
-        );
-      },
-    [modal, group]
-  );
-};
-
-const useClearFiltered = (modal: boolean, group: string) => {
-  return useRecoilCallback(
-    ({ set, snapshot }) =>
-      async () => {
-        const paths = await snapshot.getPromise(
-          fos.sidebarGroup({ modal, group, loading: true })
-        );
-        const filters = await snapshot.getPromise(
-          modal ? fos.modalFilters : fos.filters
-        );
-
-        set(
-          modal ? fos.modalFilters : fos.filters,
-          removeKeys(filters, paths, true)
-        );
-      },
-    [modal, group]
-  );
-};
-
-const useClearVisibility = (modal: boolean, group: string) => {
-  return useRecoilCallback(
-    ({ set, snapshot }) =>
-      async () => {
-        const paths = await snapshot.getPromise(
-          fos.sidebarGroup({ modal, group, loading: true })
-        );
-        const visibility = await snapshot.getPromise(
-          modal ? modalAttributeVisibility : attributeVisibility
-        );
-
-        set(
-          modal ? modalAttributeVisibility : attributeVisibility,
-          removeKeys(visibility, paths, true)
-        );
-      },
-    [modal, group]
-  );
-};
 
 type PillEntry = {
   onClick: () => void;
